pyproject.toml
[project]
name = "stduycafe"
version = "0.1.0"
description = ""
authors = [
    {name = "jhp0714",email = "gowns171@naver.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=4.2,<5.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
